Add "--diff-only" option
This shows differences without deploying the new version. This is useful
to run interactively (e.g. when modifying the config), but I'm also
planning to make this run each day via Jenkins and email us if there are
changes.

This way people can make testing changes via the Marathon UI, but we get
notified to eventually correct them if they forget to incorporate/revert
them.